Diagram 1: single host with Fabric Manager (FM)
Jonathan Cameron: Whilst accurate that you might only be able
to control the switch, I would add one MLD to this
diagram so that tunneling can be discussed.  It's minimal in
sense of near minimal number of components that might exist
in a switched system.

Reworked as suggested by Jonathan Cameron.

Signed-off-by: Viacheslav Dubeyko <viacheslav.dubeyko@bytedance.com>