Diagram 2: single host with orchestrator
Jonathan Cameron: Actually having an FM in a switch might happen,
but there is no spec defined way of doing it. From a software
architecture point of view it's no different from another host
talking to the switch - just think of sticking a BMC down next
to the switch chip. Having the orchestrator in the host is also
rather odd though it could in theory happen. Typically the orchestrator
is considered a 'cloud' level thing floating way above individual host.
Without any loss of generality I'd always have the orchestrator as
something on it's own machine chatting over a network to the Agents
and FM-API accessed devices.

Reworked as suggested by Jonathan Cameron.

Signed-off-by: Viacheslav Dubeyko <viacheslav.dubeyko@bytedance.com>