Diagram 8: BMC based layered Fabric Manager (FM)
Jonathan Cameron: The BMC is just a (cheap) host that happens
to have some elements of the overall management
infrastructure on it.  Let it be any of the FMs floating
around on their own in the diagrams above.  The diagram
immediately above might be built with 3 BMCs or as a single
BMC like this...

Reworked as requested by Jonathan Cameron.

Signed-off-by: Viacheslav Dubeyko <viacheslav.dubeyko@bytedance.com>